fixup! validatingadmissionpolicy test: use resource from the chart directory
