+++ test/helpers/apply/apply.go
@@ -7,12 +7,7 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"path"
-	"testing"
 
-	"github.com/gruntwork-io/terratest/modules/helm"
-	terratestlog "github.com/gruntwork-io/terratest/modules/logger"
-	"github.com/stretchr/testify/require"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
@@ -24,68 +19,12 @@ import (
 	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/restmapper"
-	"sigs.k8s.io/kustomize/api/krusty"
-	"sigs.k8s.io/kustomize/kyaml/filesys"
-
-	"github.com/kong/kong-operator/pkg/utils/test"
 )
 
 var decUnstructured = yaml.NewDecodingSerializer(unstructured.UnstructuredJSONScheme)
 
-// Template applies templated resources to the cluster using the given rest.Config.
-func Template(t *testing.T, cfg *rest.Config, chartPath string, templates []string) {
-	t.Helper()
-
-	helmArgs := []string{
-		"--api-versions",
-		"admissionregistration.k8s.io/v1/ValidatingAdmissionPolicy",
-		"--api-versions",
-		"admissionregistration.k8s.io/v1/ValidatingAdmissionPolicyBinding",
-	}
-
-	data := templateRenderer(t, chartPath, templates, helmArgs...)
-	res, err := apply(t.Context(), cfg, []byte(data))
-	require.NoError(t, err)
-	for _, r := range res {
-		t.Logf("Result: %s", r)
-	}
-}
-
-func templateRenderer(t *testing.T, chartPath string, templates []string, helmArgs ...string) string {
-	t.Helper()
-	releaseName := "ko"
-	valuesFile := path.Join(chartPath, "values.yaml")
-
-	// Discard terratest stdout logging
-	terratestlog.Default = terratestlog.Discard
-
-	res := helm.RenderTemplate(t, &helm.Options{
-		ValuesFiles: []string{valuesFile},
-	}, chartPath, releaseName, templates, helmArgs...)
-
-	return res
-}
-
-// Kustomization applies a kustomization to the cluster using the given rest.Config.
-func Kustomization(ctx context.Context, t *testing.T, cfg *rest.Config, dir string) {
-	t.Helper()
-
-	k := krusty.MakeKustomizer(krusty.MakeDefaultOptions())
-	fSys := filesys.MakeFsOnDisk()
-	resmap, err := k.Run(fSys, path.Join(test.ProjectRootPath(), dir))
-	require.NoError(t, err)
-
-	b, err := resmap.AsYaml()
-	require.NoError(t, err)
-
-	res, err := apply(ctx, cfg, b)
-	require.NoError(t, err)
-	for _, r := range res {
-		t.Logf("Result: %s", r)
-	}
-}
-
-func apply(ctx context.Context, restConfig *rest.Config, data []byte) (result []string, err error) {
+// Apply applies the given YAML manifest to the cluster using the provided rest.Config.
+func Apply(ctx context.Context, restConfig *rest.Config, data []byte) (result []string, err error) {
 	chanMes, chanErr := readYaml(data)
 	for {
 		select {
